Make final fixes

+++ src/main/java/app/gui/Application.java
@@ -363,7 +363,7 @@ public class Application {
              public void actionPerformed(ActionEvent e) {
                  final String course = courseField.getText();
                  try {
-                     final int top = getTopGradeUseCase.getTopGrade(course);
+                     final float top = getTopGradeUseCase.getTopGrade(course);
                      JOptionPane.showMessageDialog(jFrame, "Top Grade: " + top);
                      courseField.setText("");
                  } catch (JSONException ex) {
